bulk rename getball-intercept to getball_intercept, according to new no-dash component naming design rule, see commit b553554
