+++ components/falcons/getball/dependencies
@@ -1,2 +1,2 @@
 components/falcons/getball_fetch
-components/falcons/getball-intercept
+components/falcons/getball_intercept
