+++ components/falcons/getball/interface/Params.proto
@@ -3,7 +3,7 @@ syntax = "proto3";
 package MRA.FalconsGetball;
 
 import "components/falcons/getball_fetch/interface/Params.proto";
-import "components/falcons/getball-intercept/interface/Params.proto";
+import "components/falcons/getball_intercept/interface/Params.proto";
 
 message Params
 {
