+++ components/falcons/getball_intercept/BUILD
@@ -8,14 +8,14 @@ cc_library(
     ],
     visibility = ["//visibility:public"],
     deps = [
-        "//components/robotsports/getball-intercept/interface:interface",
+        "//components/falcons/getball_intercept/interface:interface",
     ],
 )
 
 cc_library(
     name = "implementation",
     hdrs = [
-        "RobotsportsGetballIntercept.hpp", # generated by MRA-codegen.py
+        "FalconsGetballIntercept.hpp", # generated by MRA-codegen.py
     ],
     srcs = [
         "tick.cpp", # to be implemented by user, generated if not existing
@@ -25,7 +25,7 @@ cc_library(
     deps = [
         "//base:abstract_interface",
         ":datatypes",
-        "//libraries/geometry",
+        
     ],
 )
