+++ components/robotsports/getball_intercept/BUILD
@@ -8,14 +8,14 @@ cc_library(
     ],
     visibility = ["//visibility:public"],
     deps = [
-        "//components/falcons/getball-intercept/interface:interface",
+        "//components/robotsports/getball_intercept/interface:interface",
     ],
 )
 
 cc_library(
     name = "implementation",
     hdrs = [
-        "FalconsGetballIntercept.hpp", # generated by MRA-codegen.py
+        "RobotsportsGetballIntercept.hpp", # generated by MRA-codegen.py
     ],
     srcs = [
         "tick.cpp", # to be implemented by user, generated if not existing
@@ -25,7 +25,7 @@ cc_library(
     deps = [
         "//base:abstract_interface",
         ":datatypes",
-        
+        "//libraries/geometry",
     ],
 )
