+++ components/robotsports/getball_intercept/datatypes.hpp
@@ -4,9 +4,9 @@
 #ifndef _MRA_ROBOTSPORTS_GETBALL_INTERCEPT_DATATYPES_HPP
 #define _MRA_ROBOTSPORTS_GETBALL_INTERCEPT_DATATYPES_HPP
 
-#include "components/robotsports/getball-intercept/interface/Input.pb.h"
-#include "components/robotsports/getball-intercept/interface/Params.pb.h"
-#include "components/robotsports/getball-intercept/interface/Output.pb.h"
+#include "components/robotsports/getball_intercept/interface/Input.pb.h"
+#include "components/robotsports/getball_intercept/interface/Params.pb.h"
+#include "components/robotsports/getball_intercept/interface/Output.pb.h"
 
 
 #endif
